Upload players with NULL data
Some players in the original datasets do not have attributes of birth, height, weight, team and position. I excluded them originally from the player.sql.
Now here they are with the NULL value if we have data missing for the attributes. Use this together with player.sql (or we can combine them together)